Refactor code and add new features
